backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
google-generativeai==0.8.3
pinecone-client==5.0.1
PyMuPDF==1.24.13
python-multipart==0.0.12
